+++ views/motion/_set_proposed_procedure.php
@@ -13,7 +13,6 @@ $saveUrl = UrlHelper::createMotionUrl($motion, 'save-proposal-status');
 echo Html::beginForm($saveUrl, 'POST', [
     'id'                       => 'proposedChanges',
     'data-antragsgruen-widget' => 'backend/ChangeProposedProcedure',
-    'class'                    => 'fuelux',
 ]);
 if ($motion->proposalStatus === Motion::STATUS_REFERRED) {
     $preReferredTo = $motion->proposalComment;
@@ -87,15 +86,19 @@ $selectedTags = $motion->getProposedProcedureTags();
         </fieldset>
         <div class="votingBlockSettings showIfStatusSet">
             <h3><?= Yii::t('amend', 'proposal_voteblock') ?></h3>
-            <?php
-            $options = ['-'];
-            foreach ($votingBlocks as $votingBlock) {
-                $options[$votingBlock->id] = $votingBlock->title;
-            }
-            $options['NEW'] = '- ' . Yii::t('amend', 'proposal_voteblock_newopt') . ' -';
-            $attrs          = ['id' => 'votingBlockId'];
-            echo HTMLTools::fueluxSelectbox('votingBlockId', $options, $motion->votingBlockId, $attrs);
-            ?>
+            <select name="votingBlockId" id="votingBlockId">
+                <option>-</option>
+                <?php
+                foreach ($votingBlocks as $votingBlock) {
+                    echo '<option value="' . Html::encode($votingBlock->id) . '"';
+                    if ($motion->votingBlockId === $votingBlock->id) {
+                        echo ' selected';
+                    }
+                    echo '>' . Html::encode($votingBlock->title) . '</option>';
+                }
+                ?>
+                <option value="NEW">- <?= Yii::t('amend', 'proposal_voteblock_newopt') ?> -</option>
+            </select>
             <div class="newBlock">
                 <label for="newBlockTitle" class="control-label">
                     <?= Yii::t('amend', 'proposal_voteblock_new') ?>: